requirements.txt
SpeechRecognition
pyttsx3
Pillow
